add delete code

+++ application-operator/mcagent/mcagent_secret.go
@@ -4,6 +4,9 @@
 package mcagent
 
 import (
+	"fmt"
+	"strings"
+
 	clustersv1alpha1 "github.com/verrazzano/verrazzano/application-operator/apis/clusters/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
@@ -24,14 +27,14 @@ func (s *Syncer) syncSecretObjects(namespace string) error {
 	// Write each of the secrets that are targeted for the local cluster
 	for _, mcAppConfig := range allAdminMCAppConfigs.Items {
 		if s.isThisCluster(mcAppConfig.Spec.Placement) {
-			for _, adminSecret := range mcAppConfig.Spec.Secrets{
+			for _, adminSecret := range mcAppConfig.Spec.Secrets {
 				secret := corev1.Secret{}
 				namespacedName := types.NamespacedName{Name: adminSecret, Namespace: namespace}
 				err := s.AdminClient.Get(s.Context, namespacedName, &secret)
 				if err != nil {
 					return err
 				}
-				_, err = s.createOrUpdateSecret(secret)
+				_, err = s.createOrUpdateSecret(secret, mcAppConfig.Name)
 				if err != nil {
 					s.Log.Error(err, "Error syncing object",
 						"Secret",
@@ -44,45 +47,61 @@ func (s *Syncer) syncSecretObjects(namespace string) error {
 	// Delete orphaned or no longer placed Secret resources.
 	// Get the list of Secret resources on the local cluster and compare to the list received from the admin cluster.
 	// The admin cluster is the source of truth.
-/*	allLocalSecrets := corev1.Secrets{}
+	allLocalSecrets := corev1.SecretList{}
+	listOptions = &client.ListOptions{Namespace: namespace}
 	err = s.LocalClient.List(s.Context, &allLocalSecrets, listOptions)
 	if err != nil {
 		s.Log.Error(err, "failed to list Secrets on local cluster")
 		return nil
 	}
 	for _, secret := range allLocalSecrets.Items {
-		// Delete each Secret object that is no longer placed on this cluster
+		_, found := secret.Labels[mcAppConfigsLabel]
+		// Only look at the secrets we have synced
+		if !found {
+			continue
+		}
+		// Delete each Secret object that is no longer placed on this local cluster
 		if !s.secretPlacedOnCluster(&secret, &allAdminMCAppConfigs) {
 			err := s.LocalClient.Delete(s.Context, &secret)
 			if err != nil {
-				s.Log.Error(err, fmt.Sprintf("failed to delete Secret with name %q and namespace %q", secret.Name, secret.Namespace))
+				s.Log.Error(err, fmt.Sprintf("failed to delete Secret with name %s and namespace %s", secret.Name, secret.Namespace))
 			}
 		}
 	}
-*/
+
 	return nil
 }
 
 // Create or update a Secret
-func (s *Syncer) createOrUpdateSecret(secret corev1.Secret) (controllerutil.OperationResult, error) {
+func (s *Syncer) createOrUpdateSecret(secret corev1.Secret, mcAppConfigName string) (controllerutil.OperationResult, error) {
 	var secretNew corev1.Secret
 	secretNew.Namespace = secret.Namespace
 	secretNew.Name = secret.Name
 
 	// Create or update on the local cluster
 	return controllerutil.CreateOrUpdate(s.Context, s.LocalClient, &secretNew, func() error {
-		mutateSecret(s.ManagedClusterName, secret, &secretNew)
+		mutateSecret(s.ManagedClusterName, mcAppConfigName, secret, &secretNew)
 		return nil
 	})
 }
 
 // mutateSecret mutates the Secret to reflect the contents of the parent Secret
-func mutateSecret(managedClusteName string, secret corev1.Secret, secretNew *corev1.Secret) {
+func mutateSecret(managedClusterName string, mcAppConfigName string, secret corev1.Secret, secretNew *corev1.Secret) {
 	secretNew.Labels = secret.Labels
 	if secretNew.Labels == nil {
 		secretNew.Labels = make(map[string]string)
 	}
-	secretNew.Labels["verrazzano.io/managed-cluster"] = managedClusteName
+	secretNew.Labels[managedClusterLabel] = managedClusterName
+
+	appConfigs, found := secretNew.Labels[mcAppConfigsLabel]
+	if found {
+		splitAppConfigs := strings.Split(appConfigs, ",")
+		splitAppConfigs = append(splitAppConfigs, mcAppConfigName)
+		secretNew.Labels[mcAppConfigsLabel] = strings.Join(splitAppConfigs, ",")
+
+	} else {
+		secretNew.Labels[mcAppConfigsLabel] = mcAppConfigName
+	}
 
 	secretNew.Annotations = secret.Annotations
 	secretNew.Type = secret.Type
@@ -91,13 +110,22 @@ func mutateSecret(managedClusteName string, secret corev1.Secret, secretNew *cor
 	secretNew.StringData = secret.StringData
 }
 
-// secretPlacedOnCluster returns boolean indicating if the list contains the object with the specified name and namespace
-// and indicates the object is placed on the local cluster
-func (s *Syncer) secretPlacedOnCluster(mcAdminList *clustersv1alpha1.MultiClusterSecretList, name string, namespace string) bool {
-	for _, item := range mcAdminList.Items {
-		if item.Name == name && item.Namespace == namespace {
-			return s.isThisCluster(item.Spec.Placement)
+// secretPlacedOnCluster returns boolean indicating if the secret if placed on the local cluster
+func (s *Syncer) secretPlacedOnCluster(secret *corev1.Secret, allAdminMCAppConfigs *clustersv1alpha1.MultiClusterApplicationConfigurationList) bool {
+	labels := strings.Split(secret.Labels[mcAppConfigsLabel], ",")
+	for _, mcAppConfig := range allAdminMCAppConfigs.Items {
+		for _, label := range labels {
+			// Both a matching application configuration label and a matching cluster label be found for the
+			// secret to be placed on the local cluster.
+			if mcAppConfig.Name == label {
+				for _, cluster := range mcAppConfig.Spec.Placement.Clusters {
+					if cluster.Name == secret.Labels[managedClusterLabel] {
+						return true
+					}
+				}
+			}
 		}
 	}
+
 	return false
 }
